requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
Django==4.2.4
djangorestframework==3.15.2
logging==0.4.9.6
minio==7.2.10
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
redis==5.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
numpy==2.1.2
